RedScarf/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 模型相关
ultralytics>=8.0.0
onnx>=1.14.0
openvino-dev>=2023.0.0

# 计算机视觉
opencv-python>=4.8.0
numpy>=1.24.0

# Web界面
gradio>=4.0.0

# 工具库
Pillow>=10.0.0